Move the metrics definition to pdata. Remove internal/data definition
