+++ exporter/exporterhelper/metricshelper_test.go
@@ -26,7 +26,6 @@ import (
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
 	"go.opentelemetry.io/collector/internal/data/testdata"
 	"go.opentelemetry.io/collector/obsreport"
 	"go.opentelemetry.io/collector/obsreport/obsreporttest"
@@ -46,7 +45,7 @@ var (
 )
 
 func TestMetricsRequest(t *testing.T) {
-	mr := newMetricsRequest(context.Background(), pdatautil.MetricsFromInternalMetrics(testdata.GenerateMetricsEmpty()), nil)
+	mr := newMetricsRequest(context.Background(), testdata.GenerateMetricsEmpty(), nil)
 
 	partialErr := consumererror.PartialTracesError(errors.New("some error"), testdata.GenerateTraceDataOneSpan())
 	assert.Same(t, mr, mr.onPartialError(partialErr.(consumererror.PartialError)))
@@ -71,7 +70,7 @@ func TestMetricsExporter_Default(t *testing.T) {
 	assert.NotNil(t, me)
 	assert.NoError(t, err)
 
-	assert.Nil(t, me.ConsumeMetrics(context.Background(), pdatautil.MetricsFromInternalMetrics(md)))
+	assert.Nil(t, me.ConsumeMetrics(context.Background(), md))
 	assert.Nil(t, me.Shutdown(context.Background()))
 }
 
@@ -81,7 +80,7 @@ func TestMetricsExporter_Default_ReturnError(t *testing.T) {
 	me, err := NewMetricsExporter(fakeMetricsExporterConfig, newPushMetricsData(0, want))
 	require.Nil(t, err)
 	require.NotNil(t, me)
-	require.Equal(t, want, me.ConsumeMetrics(context.Background(), pdatautil.MetricsFromInternalMetrics(md)))
+	require.Equal(t, want, me.ConsumeMetrics(context.Background(), md))
 }
 
 func TestMetricsExporter_WithRecordMetrics(t *testing.T) {
@@ -168,7 +167,7 @@ func checkRecordedMetricsForMetricsExporter(t *testing.T, me component.MetricsEx
 	md := testdata.GenerateMetricsTwoMetrics()
 	const numBatches = 7
 	for i := 0; i < numBatches; i++ {
-		require.Equal(t, wantError, me.ConsumeMetrics(context.Background(), pdatautil.MetricsFromInternalMetrics(md)))
+		require.Equal(t, wantError, me.ConsumeMetrics(context.Background(), md))
 	}
 
 	// TODO: When the new metrics correctly count partial dropped fix this.
@@ -185,7 +184,7 @@ func generateMetricsTraffic(t *testing.T, me component.MetricsExporter, numReque
 	ctx, span := trace.StartSpan(context.Background(), fakeMetricsParentSpanName, trace.WithSampler(trace.AlwaysSample()))
 	defer span.End()
 	for i := 0; i < numRequests; i++ {
-		require.Equal(t, wantError, me.ConsumeMetrics(ctx, pdatautil.MetricsFromInternalMetrics(md)))
+		require.Equal(t, wantError, me.ConsumeMetrics(ctx, md))
 	}
 }
