+++ exporter/exportertest/sink_exporter_test.go
@@ -23,7 +23,6 @@ import (
 
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
 	"go.opentelemetry.io/collector/internal/data/testdata"
 )
 
@@ -61,8 +60,8 @@ func TestSinkMetricsExporter(t *testing.T) {
 	md := testdata.GenerateMetricsOneMetric()
 	want := make([]pdata.Metrics, 0, 7)
 	for i := 0; i < 7; i++ {
-		require.NoError(t, sink.ConsumeMetrics(context.Background(), pdatautil.MetricsFromInternalMetrics(md)))
-		want = append(want, pdatautil.MetricsFromInternalMetrics(md))
+		require.NoError(t, sink.ConsumeMetrics(context.Background(), md))
+		want = append(want, md)
 	}
 	assert.Equal(t, want, sink.AllMetrics())
 	assert.Equal(t, len(want), sink.MetricsCount())
@@ -77,7 +76,7 @@ func TestSinkMetricsExporter_Error(t *testing.T) {
 	require.NoError(t, sink.Start(context.Background(), componenttest.NewNopHost()))
 	sink.SetConsumeMetricsError(errors.New("my error"))
 	md := testdata.GenerateMetricsOneMetric()
-	require.Error(t, sink.ConsumeMetrics(context.Background(), pdatautil.MetricsFromInternalMetrics(md)))
+	require.Error(t, sink.ConsumeMetrics(context.Background(), md))
 	assert.Len(t, sink.AllMetrics(), 0)
 	assert.Equal(t, 0, sink.MetricsCount())
 	require.NoError(t, sink.Shutdown(context.Background()))
