+++ exporter/otlpexporter/otlp_test.go
@@ -32,8 +32,6 @@ import (
 	"go.opentelemetry.io/collector/config/configgrpc"
 	"go.opentelemetry.io/collector/config/configtls"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
-	"go.opentelemetry.io/collector/internal/data"
 	otlplogs "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/collector/logs/v1"
 	otlpmetrics "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/collector/metrics/v1"
 	otlptraces "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/collector/trace/v1"
@@ -135,7 +133,7 @@ func (r *mockMetricsReceiver) Export(
 	req *otlpmetrics.ExportMetricsServiceRequest,
 ) (*otlpmetrics.ExportMetricsServiceResponse, error) {
 	atomic.AddInt32(&r.requestCount, 1)
-	_, recordCount := data.MetricDataFromOtlp(req.ResourceMetrics).MetricAndDataPointCount()
+	_, recordCount := pdata.MetricsFromOtlp(req.ResourceMetrics).MetricAndDataPointCount()
 	atomic.AddInt32(&r.totalItems, int32(recordCount))
 	r.lastRequest = req
 	r.metadata, _ = metadata.FromIncomingContext(ctx)
@@ -266,7 +264,7 @@ func TestSendMetrics(t *testing.T) {
 	assert.EqualValues(t, 0, atomic.LoadInt32(&rcv.requestCount))
 
 	// Send empty trace.
-	md := pdatautil.MetricsFromInternalMetrics(testdata.GenerateMetricsEmpty())
+	md := testdata.GenerateMetricsEmpty()
 	assert.NoError(t, exp.ConsumeMetrics(context.Background(), md))
 
 	// Wait until it is received.
@@ -278,7 +276,7 @@ func TestSendMetrics(t *testing.T) {
 	assert.EqualValues(t, 0, atomic.LoadInt32(&rcv.totalItems))
 
 	// A trace with 2 spans.
-	md = pdatautil.MetricsFromInternalMetrics(testdata.GenerateMetricsTwoMetrics())
+	md = testdata.GenerateMetricsTwoMetrics()
 
 	expectedOTLPReq := &otlpmetrics.ExportMetricsServiceRequest{
 		ResourceMetrics: testdata.GenerateMetricsOtlpTwoMetrics(),
