+++ exporter/prometheusremotewriteexporter/exporter_test.go
@@ -23,7 +23,7 @@ import (
 	"sync"
 	"testing"
 
-	proto "github.com/gogo/protobuf/proto"
+	"github.com/gogo/protobuf/proto"
 	"github.com/golang/snappy"
 	"github.com/prometheus/prometheus/prompb"
 	"github.com/stretchr/testify/assert"
@@ -34,7 +34,6 @@ import (
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/consumer/pdatautil"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
-	"go.opentelemetry.io/collector/internal/data"
 	otlp "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/metrics/v1"
 	"go.opentelemetry.io/collector/internal/data/testdata"
 )
@@ -115,8 +114,7 @@ func Test_Shutdown(t *testing.T) {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			_, ok := prwe.PushMetrics(context.Background(),
-				pdatautil.MetricsFromInternalMetrics(testdata.GenerateMetricsEmpty()))
+			_, ok := prwe.PushMetrics(context.Background(), testdata.GenerateMetricsEmpty())
 			errChan <- ok
 		}()
 	}
@@ -234,27 +232,26 @@ func runExportPipeline(t *testing.T, ts *prompb.TimeSeries, endpoint *url.URL) e
 // expected
 func Test_PushMetrics(t *testing.T) {
 
-	invalidTypeBatch := pdatautil.MetricsFromInternalMetrics((testdata.GenerateMetricsMetricTypeInvalid()))
+	invalidTypeBatch := testdata.GenerateMetricsMetricTypeInvalid()
 
 	nilBatch1 := testdata.GenerateMetricsOneEmptyResourceMetrics()
 	nilBatch2 := testdata.GenerateMetricsOneEmptyInstrumentationLibrary()
 	nilBatch3 := testdata.GenerateMetricsOneMetric()
 
-	nilResource := data.MetricDataToOtlp(nilBatch1)
+	nilResource := pdata.MetricsToOtlp(nilBatch1)
 	nilResource[0] = nil
-	nilResourceBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(nilResource))
+	nilResourceBatch := pdata.MetricsFromOtlp(nilResource)
 
-	nilInstrumentation := data.MetricDataToOtlp(nilBatch2)
+	nilInstrumentation := pdata.MetricsToOtlp(nilBatch2)
 	nilInstrumentation[0].InstrumentationLibraryMetrics[0] = nil
-	nilInstrumentationBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(nilInstrumentation))
+	nilInstrumentationBatch := pdata.MetricsFromOtlp(nilInstrumentation)
 
-	nilMetric := data.MetricDataToOtlp(nilBatch3)
+	nilMetric := pdata.MetricsToOtlp(nilBatch3)
 	nilMetric[0].InstrumentationLibraryMetrics[0].Metrics[0] = nil
-	nilMetricBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(nilMetric))
+	nilMetricBatch := pdata.MetricsFromOtlp(nilMetric)
 
 	// success cases
-	intSumMetric := testdata.GenerateMetricsManyMetricsSameResource(10)
-	intSumBatch := pdatautil.MetricsFromInternalMetrics(intSumMetric)
+	intSumBatch := testdata.GenerateMetricsManyMetricsSameResource(10)
 
 	doubleSumMetric := []*otlp.ResourceMetrics{
 		{
@@ -268,7 +265,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	doubleSumBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(doubleSumMetric))
+	doubleSumBatch := pdata.MetricsFromOtlp(doubleSumMetric)
 
 	intGaugeMetric := []*otlp.ResourceMetrics{
 		{
@@ -282,7 +279,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	intGaugeBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(intGaugeMetric))
+	intGaugeBatch := pdata.MetricsFromOtlp(intGaugeMetric)
 
 	doubleGaugeMetric := []*otlp.ResourceMetrics{
 		{
@@ -296,7 +293,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	doubleGaugeBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(doubleGaugeMetric))
+	doubleGaugeBatch := pdata.MetricsFromOtlp(doubleGaugeMetric)
 
 	intHistogramMetric := []*otlp.ResourceMetrics{
 		{
@@ -310,7 +307,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	intHistogramBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(intHistogramMetric))
+	intHistogramBatch := pdata.MetricsFromOtlp(intHistogramMetric)
 
 	doubleHistogramMetric := []*otlp.ResourceMetrics{
 		{
@@ -324,7 +321,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	doubleHistogramBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(doubleHistogramMetric))
+	doubleHistogramBatch := pdata.MetricsFromOtlp(doubleHistogramMetric)
 
 	// len(BucketCount) > len(ExplicitBounds)
 	unmatchedBoundBucketIntHistMetric := []*otlp.ResourceMetrics{
@@ -338,8 +335,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	unmatchedBoundBucketIntHistBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(unmatchedBoundBucketIntHistMetric))
+	unmatchedBoundBucketIntHistBatch := pdata.MetricsFromOtlp(unmatchedBoundBucketIntHistMetric)
 
 	unmatchedBoundBucketDoubleHistMetric := []*otlp.ResourceMetrics{
 		{
@@ -352,8 +348,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	unmatchedBoundBucketDoubleHistBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(unmatchedBoundBucketDoubleHistMetric))
+	unmatchedBoundBucketDoubleHistBatch := pdata.MetricsFromOtlp(unmatchedBoundBucketDoubleHistMetric)
 
 	// fail cases
 	nilDataPointIntGaugeMetric := []*otlp.ResourceMetrics{
@@ -367,8 +362,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointIntGaugeBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(nilDataPointIntGaugeMetric))
+	nilDataPointIntGaugeBatch := pdata.MetricsFromOtlp(nilDataPointIntGaugeMetric)
 
 	nilDataPointDoubleGaugeMetric := []*otlp.ResourceMetrics{
 		{
@@ -381,8 +375,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointDoubleGaugeBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(nilDataPointDoubleGaugeMetric))
+	nilDataPointDoubleGaugeBatch := pdata.MetricsFromOtlp(nilDataPointDoubleGaugeMetric)
 
 	nilDataPointIntSumMetric := []*otlp.ResourceMetrics{
 		{
@@ -395,7 +388,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointIntSumBatch := pdatautil.MetricsFromInternalMetrics(data.MetricDataFromOtlp(nilDataPointIntSumMetric))
+	nilDataPointIntSumBatch := pdata.MetricsFromOtlp(nilDataPointIntSumMetric)
 
 	nilDataPointDoubleSumMetric := []*otlp.ResourceMetrics{
 		{
@@ -408,8 +401,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointDoubleSumBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(nilDataPointDoubleSumMetric))
+	nilDataPointDoubleSumBatch := pdata.MetricsFromOtlp(nilDataPointDoubleSumMetric)
 
 	nilDataPointIntHistogramMetric := []*otlp.ResourceMetrics{
 		{
@@ -422,8 +414,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointIntHistogramBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(nilDataPointIntHistogramMetric))
+	nilDataPointIntHistogramBatch := pdata.MetricsFromOtlp(nilDataPointIntHistogramMetric)
 
 	nilDataPointDoubleHistogramMetric := []*otlp.ResourceMetrics{
 		{
@@ -436,8 +427,7 @@ func Test_PushMetrics(t *testing.T) {
 			},
 		},
 	}
-	nilDataPointDoubleHistogramBatch := pdatautil.MetricsFromInternalMetrics(
-		data.MetricDataFromOtlp(nilDataPointDoubleHistogramMetric))
+	nilDataPointDoubleHistogramBatch := pdata.MetricsFromOtlp(nilDataPointDoubleHistogramMetric)
 
 	checkFunc := func(t *testing.T, r *http.Request, expected int) {
 		body, err := ioutil.ReadAll(r.Body)
