+++ internal/data/testdata/metric_test.go
@@ -19,13 +19,13 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/collector/internal/data"
+	"go.opentelemetry.io/collector/consumer/pdata"
 	otlpmetrics "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/metrics/v1"
 )
 
 type traceMetricsCase struct {
 	name string
-	td   data.MetricData
+	td   pdata.Metrics
 	otlp []*otlpmetrics.ResourceMetrics
 }
 
@@ -111,9 +111,9 @@ func TestToFromOtlpMetrics(t *testing.T) {
 	for i := range allTestCases {
 		test := allTestCases[i]
 		t.Run(test.name, func(t *testing.T) {
-			td := data.MetricDataFromOtlp(test.otlp)
+			td := pdata.MetricsFromOtlp(test.otlp)
 			assert.EqualValues(t, test.td, td)
-			otlp := data.MetricDataToOtlp(td)
+			otlp := pdata.MetricsToOtlp(td)
 			assert.EqualValues(t, test.otlp, otlp)
 		})
 	}
