+++ processor/resourceprocessor/resource_processor_test.go
@@ -24,7 +24,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
 	"go.opentelemetry.io/collector/internal/data/testdata"
 	"go.opentelemetry.io/collector/processor/processorhelper"
 )
@@ -148,7 +147,7 @@ func generateTraceData(attributes map[string]string) pdata.Traces {
 func generateMetricData(attributes map[string]string) pdata.Metrics {
 	md := testdata.GenerateMetricsOneMetricNoResource()
 	if attributes == nil {
-		return pdatautil.MetricsFromInternalMetrics(md)
+		return md
 	}
 	resource := md.ResourceMetrics().At(0).Resource()
 	resource.InitEmpty()
@@ -156,7 +155,7 @@ func generateMetricData(attributes map[string]string) pdata.Metrics {
 		resource.Attributes().InsertString(k, v)
 	}
 	resource.Attributes().Sort()
-	return pdatautil.MetricsFromInternalMetrics(md)
+	return md
 }
 
 type testTraceConsumer struct {
@@ -178,9 +177,8 @@ type testMetricsConsumer struct {
 
 func (tmn *testMetricsConsumer) ConsumeMetrics(_ context.Context, md pdata.Metrics) error {
 	// sort attributes to be able to compare traces
-	imd := pdatautil.MetricsToInternalMetrics(md)
-	for i := 0; i < imd.ResourceMetrics().Len(); i++ {
-		sortResourceAttributes(imd.ResourceMetrics().At(i).Resource())
+	for i := 0; i < md.ResourceMetrics().Len(); i++ {
+		sortResourceAttributes(md.ResourceMetrics().At(i).Resource())
 	}
 	tmn.md = md
 	return nil
