+++ receiver/hostmetricsreceiver/hostmetrics_receiver.go
@@ -25,8 +25,7 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenterror"
 	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
-	"go.opentelemetry.io/collector/internal/data"
+	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver/internal"
 )
 
@@ -162,7 +161,7 @@ func (hmr *receiver) scrapeMetrics(ctx context.Context) {
 	defer span.End()
 
 	var errors []error
-	metricData := data.NewMetricData()
+	metricData := pdata.NewMetrics()
 
 	if err := hmr.scrapeAndAppendHostMetrics(ctx, metricData); err != nil {
 		errors = append(errors, err)
@@ -176,13 +175,13 @@ func (hmr *receiver) scrapeMetrics(ctx context.Context) {
 		span.SetStatus(trace.Status{Code: trace.StatusCodeDataLoss, Message: fmt.Sprintf("Error(s) when scraping metrics: %v", componenterror.CombineErrors(errors))})
 	}
 
-	if err := hmr.consumer.ConsumeMetrics(ctx, pdatautil.MetricsFromInternalMetrics(metricData)); err != nil {
+	if err := hmr.consumer.ConsumeMetrics(ctx, metricData); err != nil {
 		span.SetStatus(trace.Status{Code: trace.StatusCodeDataLoss, Message: fmt.Sprintf("Unable to process metrics: %v", err)})
 		return
 	}
 }
 
-func (hmr *receiver) scrapeAndAppendHostMetrics(ctx context.Context, metricData data.MetricData) error {
+func (hmr *receiver) scrapeAndAppendHostMetrics(ctx context.Context, metricData pdata.Metrics) error {
 	if len(hmr.hostMetricScrapers) == 0 {
 		return nil
 	}
@@ -202,7 +201,7 @@ func (hmr *receiver) scrapeAndAppendHostMetrics(ctx context.Context, metricData 
 	return componenterror.CombineErrors(errors)
 }
 
-func (hmr *receiver) scrapeAndAppendResourceMetrics(ctx context.Context, metricData data.MetricData) error {
+func (hmr *receiver) scrapeAndAppendResourceMetrics(ctx context.Context, metricData pdata.Metrics) error {
 	if len(hmr.resourceMetricScrapers) == 0 {
 		return nil
 	}
