+++ testbed/testbed/data_providers.go
@@ -27,8 +27,6 @@ import (
 	"go.uber.org/atomic"
 
 	"go.opentelemetry.io/collector/consumer/pdata"
-	"go.opentelemetry.io/collector/consumer/pdatautil"
-	"go.opentelemetry.io/collector/internal/data"
 	otlptrace "go.opentelemetry.io/collector/internal/data/opentelemetry-proto-gen/trace/v1"
 	"go.opentelemetry.io/collector/internal/goldendataset"
 )
@@ -123,17 +121,17 @@ func (dp *PerfTestDataProvider) GenerateMetrics() (pdata.Metrics, bool) {
 	// Generate 7 data points per metric.
 	const dataPointsPerMetric = 7
 
-	metricData := data.NewMetricData()
-	metricData.ResourceMetrics().Resize(1)
-	metricData.ResourceMetrics().At(0).InstrumentationLibraryMetrics().Resize(1)
+	md := pdata.NewMetrics()
+	md.ResourceMetrics().Resize(1)
+	md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().Resize(1)
 	if dp.options.Attributes != nil {
-		attrs := metricData.ResourceMetrics().At(0).Resource().Attributes()
+		attrs := md.ResourceMetrics().At(0).Resource().Attributes()
 		attrs.InitEmptyWithCapacity(len(dp.options.Attributes))
 		for k, v := range dp.options.Attributes {
 			attrs.UpsertString(k, v)
 		}
 	}
-	metrics := metricData.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics()
+	metrics := md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics()
 	metrics.Resize(dp.options.ItemsPerBatch)
 
 	for i := 0; i < dp.options.ItemsPerBatch; i++ {
@@ -160,7 +158,7 @@ func (dp *PerfTestDataProvider) GenerateMetrics() (pdata.Metrics, bool) {
 			})
 		}
 	}
-	return pdatautil.MetricsFromInternalMetrics(metricData), false
+	return md, false
 }
 
 func (dp *PerfTestDataProvider) GetGeneratedSpan([]byte, []byte) *otlptrace.Span {
@@ -262,11 +260,11 @@ func (dp *GoldenDataProvider) GenerateTraces() (pdata.Traces, bool) {
 }
 
 func (dp *GoldenDataProvider) GenerateMetrics() (pdata.Metrics, bool) {
-	return pdatautil.MetricsFromInternalMetrics(data.MetricData{}), true
+	return pdata.NewMetrics(), true
 }
 
 func (dp *GoldenDataProvider) GenerateLogs() (pdata.Logs, bool) {
-	return pdata.Logs{}, true
+	return pdata.NewLogs(), true
 }
 
 func (dp *GoldenDataProvider) GetGeneratedSpan(traceID []byte, spanID []byte) *otlptrace.Span {
